Move web app example data types into `model` module.
