+++ examples/web_app/src/model/repo_slug_error.rs
