+++ examples/web_app/src/model/web_app_error.rs
@@ -10,7 +10,7 @@ pub enum WebAppError {
     EnvTypeParseError(
         #[source]
         #[from]
-        crate::EnvTypeParseError,
+        crate::model::EnvTypeParseError,
     ),
 
     // === Item Spec errors === //
